% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_penalty_table.R
\name{create_penalty_table}
\alias{create_penalty_table}
\title{Function to create table of penalties}
\usage{
create_penalty_table(snps_assoc, map, ind.col, w2, w3)
}
\arguments{
\item{snps_assoc}{output of the extract_snp_groups function: a named table where the names are SNPs and values consist of whether SNP is pheno only, endo only, or associated with both.}

\item{map}{data frame containing information about SNPs in G (genotype matrix)}

\item{ind.col}{index of SNPs in map that are in snp_assoc table}

\item{w2}{weight to assign SNPs in endo only group}

\item{w3}{weight to assign SNPs associated with both pheno and endo}
}
\value{
a data frame that contains the name of the SNP, the group (pheno only, endo only, or both), and the associated penalty to be used in the weighted lasso model.
}
\description{
This function creates a data frame that contains the name of the SNP,
the group (pheno only, endo only, or both), and the associated penalty
to be used in the weighted lasso model.
}
