% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_snp_groups.R
\name{extract_snp_groups}
\alias{extract_snp_groups}
\title{Function to extract snps associated with phenotype, with endotype, and with both}
\usage{
extract_snp_groups(pheno_gwas, endo_gwas, map, thresh, filter_hapmap, hapmap)
}
\arguments{
\item{pheno_gwas}{data frame containing results of phenotype GWAS. It should contain a column with chromosome, SNP ID, allele1, allele2, and P-value.}

\item{endo_gwas}{data frame containing results of endophenotype GWAS. It should contain a column with chromosome, SNP ID, allele1, allele2, and P-value.}

\item{map}{data frame containing information about SNPs in G (genotype matrix)}

\item{thresh}{P-value threshold used to determine SNP association with phenotype and endophenotype in GWAS.}

\item{filter_hapmap}{logical value as to whether to filter to SNPs in hapmap.}

\item{hapmap}{data frame with hapmap variants. Only provide if filter_hapmap argument set to TRUE. It should contain columns CHR, SNP, A1, A2.}
}
\value{
A named table. The name corresponds to the SNP and value corresponds to the assigned group (pheno only, endo only, or both). It is used as input in the create_penalty_table function
}
\description{
This function creates a table that contains the name of the SNP and the corresponding
group it is assigned to based on GWAS p-values (pheno only, endo only, or both). It is
used to determine penalties for the weighted lasso model.
}
