% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_endoPRS_single_iter.R
\name{fit_endoPRS_single_iter}
\alias{fit_endoPRS_single_iter}
\title{Function to apply the first step of the endoPRS method to generate a polygenic risk score model for a particular penalty.}
\usage{
fit_endoPRS_single_iter(
  G,
  map,
  fam,
  train_pheno,
  train_covar,
  val_pheno,
  val_covar,
  pheno_gwas,
  endo_gwas,
  filter_hapmap = F,
  hapmap = NULL,
  type = NULL,
  thresh,
  grid = NULL,
  iteration,
  NCORES = NULL,
  save_folder,
  save_models = F
)
}
\arguments{
\item{G}{An object of class FBM from the bigsnpr package that contains genotypes of the individuals used for training and those used for validation.}

\item{map}{A data frame containing information about SNPs in G (genotype matrix). This can be the map object form the bigSNP class. It should contain columns labeled chromosome, marker ID, and allele1 and allele2.}

\item{fam}{A data frame containing information about the individuals in G (genotype matrix). Column one must correspond to FID and column 2 must correspond to IID.}

\item{train_pheno}{A data frame with three columns corresponding to the phenotypes of individuals used for training. Column one is FID, column two is IID, and column three is y the phenotype. If the phenotype is binary, y must consist of 0's and 1's.}

\item{train_covar}{A data frame corresponding with the covariates of the individuals used for training (ie genetic PC's, sex, age, etc). The first column must be FID, and the second column must be IID. The order of individuals must be the same as train_pheno. Covariates that are factors (i.e. assessment center) must be encoded as such. The covariates are not penalized in the endoPRS model.}

\item{val_pheno}{A data frame with three columns corresponding to the phenotypes of individuals used for validation. Column one is FID, column two is IID, and column three is y the phenotype. If the phenotype is binary, y must consist of 0's and 1's.}

\item{val_covar}{A data frame corresponding with the covariates of the individuals used for validation (ie genetic PC's, sex, age, etc). The first column must be FID, and the second column must be IID. The order of individuals must be the same as val_pheno. Covariates that are factors (i.e. assessment center) must be encoded as such. The covariates are not penalized in the endoPRS model.}

\item{pheno_gwas}{A data frame containing the results of the GWAS run on the phenotype. It must not include any individuals from the validation set. It must contain columns corresponding to chromosome, SNP, allele1, allele2, and P-value.}

\item{endo_gwas}{A data frame containing the results of the GWAS run on the endophenotype. It must not include any individuals from the validation set. It must contain columns corresponding to chromosome, SNP, allele1, allele2, and P-value.}

\item{filter_hapmap}{An optional logical character. It corresponds to whether to only run the model on variants in the hapmap3 set.}

\item{hapmap}{An optional data frame with hapmap variants. This must be provided if filter_hapmap is set to TRUE. The hapmap3 data frame should contain columns CHR, SNP, A1, A2.}

\item{type}{An optional character vector of "linear" or "logistic." It specifies which type of model to fit to the data. If not provided, it will be learned from the phenotype. If the phenotype consists of 0's and 1's penalized logistic regression will be used, otherwie penalized linear regression will be used.}

\item{thresh}{An numeric vector corresponding to the p-value thresholds. This corresponds to the GWAS p-value threshold used to determine association with the phenotype and endophenotype.}

\item{grid}{An optional data frame of weights to use for the weighted penalty in the lasso model. The first column must be w2 and correspond to the weights applied to the group of SNPs associated with only the endophenotype and the second column must be w3 and correspond to the weights applied to the group of SNPs associated with both the phenotype and endophenotype. The set of SNPs associated with only the phenotype is given a weight of 1. If not provided, the original grid of weights from the endoPRS manuscript (0.1, 0.5, 1, 2, 10) will be used for both w2 and w3.}

\item{iteration}{A numerical value corresponding to which entry of the grid to fit. For example, 3 corresponds to fitting weights provided in the third row of the grid.}

\item{NCORES}{An optional value corresponding to the number of cores to use. Otherwise, the number of cores will be learned using nb_cores().}

\item{save_folder}{A path to a directory that files can be written to. The fitted model and its performance in the validation set will be saved to that directory.}

\item{save_models}{An optional boolean value. Indicates whether to save the intermediate models to disk. If set to true, the save_folder will be used.}
}
\value{
A list with two elements: the model itself and the performance in the validation set
\itemize{
\item{model: The final fitted endoPRS model as an object of class big_sp_list. This can be used to apply the final model to a test set using the predict() function from the bigsnpr package.  }
\item{val_perf: A data frame consisting of the performance of the model in the validation set. R^2 is used for linear models and AUC is used for logistic models. }
}
}
\description{
This function applies the first part of the endoPRS model. It runs a weighted lasso model that penalizes the SNPs differently based on
whether they are  associated with only the phenotype, only the endophenotype, or both. It fits only one set of penalties. This allows for
more efficient parallelization of the fit_endoPRS function. Must be used in combination with refit_endoPRS.
}
