% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weighted_model.R
\name{fit_weighted_model}
\alias{fit_weighted_model}
\title{Function to fit weighted lasso model of endoPRS}
\usage{
fit_weighted_model(
  G,
  y.train,
  train_index,
  ind.col,
  penalty_table,
  covar.train,
  NCORES,
  type
)
}
\arguments{
\item{G}{An object of class FBM from the bigsnpr package that contains genotypes.}

\item{y.train}{Vector of responses, correspoinding to train_index}

\item{train_index}{A vector of row indices that are used for training.}

\item{ind.col}{The vector of columns (SNPs) to use for fitting the model.}

\item{penalty_table}{The output of create_penalty_table a data frame that contains the name of the SNP, the group (pheno only, endo only, or both), and the associated penalty to be used in the weighted lasso model.}

\item{covar.train}{Matrix of covariates included in the model. They are unpenalized. They should correspond to train_index.}

\item{NCORES}{Number of cores to use for fitting the model.}

\item{type}{A character vector of "linear" or "logistic." It specifies which type of model to fit to the data.}
}
\value{
An object of class big_sp_list (a list of of 5) that contains the fitted model for each of the cross-validation splits.
}
\description{
This function its a weighted lasso model. It penalizes the SNPs associated with only the phenotype, only the
endophenotype, and those associated with both differently. The model is either a penalized linear or penalized
logistic model.
}
