% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gwas.R
\name{format_gwas}
\alias{format_gwas}
\title{Function to clean and format GWAS results}
\usage{
format_gwas(df, map, filter_hapmap, hapmap)
}
\arguments{
\item{df}{data frame containing results of GWAS. It should contain a column with chromosome, SNP, allele1, allele2, and P-value.}

\item{map}{data frame containing information about SNPs in G (genotype matrix)}

\item{filter_hapmap}{logical value as to whether to filter to SNPs in hapmap.}

\item{hapmap}{data frame with hapmap variants. Only provide if filter_hapmap argument set to TRUE. It should contain columns CHR, SNP, A1, A2.}
}
\value{
A data frame consisting of GWAS results that include columns CHR, SNP, A0, A1, P and a unique id. It only contains SNPs that are present in the genotype data.
}
\description{
This function cleans a dataframe of GWAS results. It formats the column names to be readable by later functions. It adds
a unique id (combination of chromosome, SNP ID, allele1, and allele2). It keeps only the SNPs present in the
genotype data (those in the map data frame). And it can also filter to keep only hapmap variants.
}
