% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_beta.R
\name{obtain_beta}
\alias{obtain_beta}
\title{Function to obtain beta from fitted lasso model.}
\usage{
obtain_beta(model, ind.col, map)
}
\arguments{
\item{model}{a big_spLinReg/big_spLogReg model created by the fit_weighted_model function.}

\item{ind.col}{a vector containing the indices of SNPs from G (genotype matrix) included as input in the fit_weighted_model.}

\item{map}{data frame containing information about SNPs in G (genotype matrix).}
}
\value{
A data frame consisting of the SNPs included in the final model and their corresponding coefficients.
}
\description{
This function takes as input the fitted model from fit_weighted_model. It extracts the coefficients of the model
estimated in all five folds and averages them. Only non zero coefficients are returned.
}
