% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_endoPRS.R
\name{refit_endoPRS}
\alias{refit_endoPRS}
\title{Function to apply the second (refitting) step of the endoPRS method to generate a polygenic risk score model.}
\usage{
refit_endoPRS(
  G,
  map,
  fam,
  train_pheno,
  train_covar,
  val_pheno,
  val_covar,
  pheno_gwas,
  endo_gwas,
  filter_hapmap = F,
  hapmap = NULL,
  type = NULL,
  threshes = c(0.01, 1e-04, 1e-06),
  NCORES = NULL,
  save_folder,
  save_model = F
)
}
\arguments{
\item{G}{An object of class FBM from the bigsnpr package that contains genotypes of the individuals used for training and those used for validation.}

\item{map}{A data frame containing information about SNPs in G (genotype matrix). This can be the map object form the bigSNP class. It should contain columns labeled chromosome, marker ID, and allele1 and allele2.}

\item{fam}{A data frame containing information about the individuals in G (genotype matrix). Column one must correspond to FID and column 2 must correspond to IID.}

\item{train_pheno}{A data frame with three columns corresponding to the phenotypes of individuals used for training. Column one is FID, column two is IID, and column three is y the phenotype. If the phenotype is binary, y must consist of 0's and 1's.}

\item{train_covar}{A data frame corresponding with the covariates of the individuals used for training (ie genetic PC's, sex, age, etc). The first column must be FID, and the second column must be IID. The order of individuals must be the same as train_pheno. Covariates that are factors (i.e. assessment center) must be encoded as such. The covariates are not penalized in the endoPRS model.}

\item{val_pheno}{A data frame with three columns corresponding to the phenotypes of individuals used for validation. Column one is FID, column two is IID, and column three is y the phenotype. If the phenotype is binary, y must consist of 0's and 1's.}

\item{val_covar}{A data frame corresponding with the covariates of the individuals used for validation (ie genetic PC's, sex, age, etc). The first column must be FID, and the second column must be IID. The order of individuals must be the same as val_pheno. Covariates that are factors (i.e. assessment center) must be encoded as such. The covariates are not penalized in the endoPRS model.}

\item{pheno_gwas}{A data frame containing the results of the GWAS run on the phenotype. It can be different than the one used in fit_endoPRS_single_iter as it can be run on the combined training and validation set. It must contain columns corresponding to chromosome, SNP, allele1, allele2, and P-value.}

\item{endo_gwas}{A data frame containing the results of the GWAS run on the endophenotype. It can be different than the one used in fit_endoPRS_single_iter as it can be run on the combined training and validation set. It must contain columns corresponding to chromosome, SNP, allele1, allele2, and P-value.}

\item{filter_hapmap}{An optional logical character. It corresponds to whether to only run the model on variants in the hapmap3 set.}

\item{hapmap}{An optional data frame with hapmap variants. This must be provided if filter_hapmap is set to TRUE. The hapmap3 data frame should contain columns CHR, SNP, A1, A2.}

\item{type}{An optional character vector of "linear" or "logistic." It specifies which type of model to fit to the data. If not provided, it will be learned from the phenotype. If the phenotype consists of 0's and 1's penalized logistic regression will be used, otherwie penalized linear regression will be used.}

\item{threshes}{An optional character of p-value thresholds. This corresponds to the GWAS p-value thresholds used to determine association with the phenotype and endophenotype. The optimal p-value threshold is learned from tuning on the validation set. If not provided, the original values from the endoPRS manuscript will be used (0.01, 1e-4, and 1e-6).}

\item{NCORES}{An optional value corresponding to the number of cores to use. Otherwise, the number of cores will be learned using nb_cores().}

\item{save_folder}{The path to a directory that contains the results of fit_endoPRS_single_iter when applied to the validation set.}

\item{save_model}{An optional boolean value. Indicates whether to save the final model and beta values to disk. If set to true, the save_folder will be used.}
}
\value{
A list with two elements: the betas of the final model and the model itself
\itemize{
\item{beta: A data frame consisting of the SNPs included in the final model and their corresponding coefficients. This can be used to apply the PRS to an external data set using software such as PLINK.}
\item{model: The final fitted endoPRS model as an object of class big_sp_list. This can be used to apply the final model to a test set using the predict() function from the bigsnpr package.  }
}
}
\description{
This function applies the refitting step of the endoPRS model. It loads the results of fit_endoPRS_single_iter to determine
the optimal chosen weights and GWAS p-value threshold in the validation set. It then refits the endoPRS model using these tuning
parameters on the combined training and validation set to obtain the final PRS model. Must be used in combination with fit_endoPRS_single_iter.
}
